Fix dpi problem
The dimensions of compute_images_mse function's parameters source_image and new_image were different on my machine. After some inspection, I found the problem was with dpi. The default dpi in matplotlib is 100. However in the code, the dpi is declared as 80 on line 91. Explicitly writing the dpi on the subplot on line 139 fixes this problem.

* https://matplotlib.org/3.2.1/api/_as_gen/matplotlib.pyplot.figure.html#matplotlib-pyplot-figure